refactor(loop-support): isolate context window cache
